requirements.txt
torch
torchvision
streamlit
Pillow
kaggle